+++ docs/src/pages/404.vue
@@ -1,0 +1,15 @@
+<template>
+  <div>
+    <h1>404: NOT FOUND</h1>
+  </div>
+</template>
+
+<script>
+export default {
+  metaInfo: {
+    title: "404: Not Found",
+  },
+}
+</script>
+
+<style lang="scss" scoped></style>